requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
cryptography==41.0.7
email-validator==2.1.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
colorama==0.4.6
openpyxl==3.1.2
